#1654 moving olog clients and module to restructured parent project
