+++ app/logbook/olog/client-es/src/main/java/phoebus/olog/es/api/XmlLog.java
