+++ app/logbook/olog/client-es/src/main/resources/META-INF/services/org.phoebus.logbook.LogFactory
