+++ app/logbook/olog/client-es/src/main/resources/olog_es_preferences.properties
