+++ app/logbook/olog/client/src/main/java/phoebus/olog/api/XmlLogs.java
