+++ app/logbook/olog/client/src/main/resources/olog_preferences.properties
